refactor: 调整包路径和例子信息
