+++ springgear-core/src/main/java/org/springgear/EnableSpringGear.java
@@ -2,18 +2,18 @@ package org.springgear;
 
 import org.springframework.context.annotation.Import;
 import org.springframework.core.annotation.AliasFor;
-import org.springgear.beans.AbstractSpringGearProxyProcessor;
-import org.springgear.beans.DefaultBeanDefinitionProcessor;
-import org.springgear.beans.annotation.SpringGearBeanRegistrarImport;
-import org.springgear.core.AbstractSpringGearEngineExecutor;
-import org.springgear.core.DefaultSpringGearEngineExecutor;
+import org.springgear.core.beans.AbstractSpringGearProxyProcessor;
+import org.springgear.core.beans.DefaultBeanDefinitionProcessor;
+import org.springgear.core.register.SpringGearBeanRegistrarAnnotation;
+import org.springgear.core.execute.AbstractSpringGearEngineExecutor;
+import org.springgear.core.execute.DefaultSpringGearEngineExecutor;
 
 import java.lang.annotation.*;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Import({SpringGearConfiguration.class, SpringGearBeanRegistrarImport.class})
+@Import({SpringGearConfiguration.class, SpringGearBeanRegistrarAnnotation.class})
 public @interface EnableSpringGear {
 
     /**
