+++ springgear-core/src/main/java/org/springgear/core/beans/SpringGearProxyInstance.java
@@ -1,9 +1,9 @@
-package org.springgear.beans;
+package org.springgear.core.beans;
 
-import org.springgear.beans.interceptor.SpringGearInterceptor;
-import org.springgear.beans.interceptor.SpringGearInterceptorChain;
-import org.springgear.context.utils.SpringGearEngineUtils;
-import org.springgear.core.SpringGearEngineInterface;
+import org.springgear.core.beans.interceptor.SpringGearInterceptor;
+import org.springgear.core.beans.interceptor.SpringGearInterceptorChain;
+import org.springgear.core.register.context.utils.SpringGearEngineUtils;
+import org.springgear.core.execute.SpringGearEngineInterface;
 import org.springgear.core.annotation.SpringGearEngine;
 import org.springgear.exception.SpringGearInterruptException;
 import org.springgear.exception.SpringGearException;
