+++ springgear-core/src/main/java/org/springgear/core/beans/interceptor/SpringGearInterceptor.java
@@ -1,4 +1,4 @@
-package org.springgear.beans.interceptor;
+package org.springgear.core.beans.interceptor;
 
 import org.springgear.exception.SpringGearException;
