+++ springgear-core/src/main/java/org/springgear/core/execute/AbstractSpringGearEngineExecutor.java
@@ -1,4 +1,4 @@
-package org.springgear.core;
+package org.springgear.core.execute;
 
 //import com.google.common.eventbus.AsyncEventBus;
 import org.springgear.core.context.EmptyRequest;
