+++ springgear-core/src/main/java/org/springgear/core/execute/SpringGearEngineInterface.java
@@ -1,4 +1,4 @@
-package org.springgear.core;
+package org.springgear.core.execute;
 
 import org.springgear.exception.SpringGearInterruptException;
 import org.springgear.exception.SpringGearException;
