+++ springgear-example/example-01-start-xml_config/src/main/java/org/springgear/example/Main.java
@@ -1,10 +1,7 @@
 package org.springgear.example;
 
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.*;
 import org.springframework.context.support.GenericXmlApplicationContext;
-import org.springgear.beans.annotation.SpringGearBeanRegister;
-import org.springgear.context.SpringGearEngineProcessor;
 import org.springgear.example.service.MyServiceInterface;
 
 public class Main {
