+++ springgear-example/example-04-simple_use/src/main/java/org/springgear/example/Main.java
@@ -3,8 +3,6 @@ package org.springgear.example;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.*;
 import org.springgear.EnableSpringGear;
-import org.springgear.beans.annotation.SpringGearBeanRegister;
-import org.springgear.context.SpringGearEngineProcessor;
 import org.springgear.example.service.MyServiceInterface;
 
 public class Main {
@@ -13,7 +11,6 @@ public class Main {
     @ComponentScan(value = "org.springgear.example")
     @EnableSpringGear
     static class Config {
-
     }
 
     public static void main(String[] args) {
@@ -33,20 +30,24 @@ public class Main {
     public static void single(MyServiceInterface service) {
         String out = service.single("this is single input params");
         System.out.println(out);
+        System.out.println("------------------------------------------------");
     }
 
     public static void multi(MyServiceInterface service) {
         String out = service.multi("this is multi input params");
         System.out.println(out);
+        System.out.println("------------------------------------------------");
     }
 
     public static void order(MyServiceInterface service) {
         String out = service.order("order");
         System.out.println(out);
+        System.out.println("------------------------------------------------");
     }
 
     public static void mixed(MyServiceInterface service) {
         String out = service.mixed("mixed");
         System.out.println(out);
+        System.out.println("------------------------------------------------");
     }
 }
